Download TIGER shapes for New England City and Town Area

Download TIGER shapes for New England City and Town Area

Usage

tt_new_england_cities(year = 2021)

Arguments

year

Integer year. Required. Years < 2012 are not currently supported.

Value

sf data.frame

Examples

tt_new_england_cities()
#> Simple feature collection with 40 features and 12 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -73.72777 ymin: 40.95094 xmax: -68.20225 ymax: 45.66082
#> Geodetic CRS:  NAD83
#> First 10 features:
#>    CNECTAFP NECTAFP GEOID                            NAME
#> 1       715   70450 70450                       Athol, MA
#> 2       710   70600 70600                     Augusta, ME
#> 3      <NA>   70750 70750                      Bangor, ME
#> 4       715   70900 70900             Barnstable Town, MA
#> 5      <NA>   71050 71050                       Barre, VT
#> 6      <NA>   71350 71350                  Bennington, VT
#> 7      <NA>   71500 71500                      Berlin, NH
#> 8       715   71650 71650  Boston-Cambridge-Newton, MA-NH
#> 9       720   71950 71950 Bridgeport-Stamford-Norwalk, CT
#> 10      775   72250 72250                   Brunswick, ME
#>                                              NAMELSAD LSAD NMEMI MTFCC
#> 1                        Athol, MA Micropolitan NECTA   M6     2 G3210
#> 2                      Augusta, ME Micropolitan NECTA   M6     2 G3210
#> 3                       Bangor, ME Metropolitan NECTA   M5     1 G3210
#> 4              Barnstable Town, MA Metropolitan NECTA   M5     1 G3210
#> 5                        Barre, VT Micropolitan NECTA   M6     2 G3210
#> 6                   Bennington, VT Micropolitan NECTA   M6     2 G3210
#> 7                       Berlin, NH Micropolitan NECTA   M6     2 G3210
#> 8   Boston-Cambridge-Newton, MA-NH Metropolitan NECTA   M5     1 G3210
#> 9  Bridgeport-Stamford-Norwalk, CT Metropolitan NECTA   M5     1 G3210
#> 10                   Brunswick, ME Micropolitan NECTA   M6     2 G3210
#>         ALAND     AWATER    INTPTLAT     INTPTLON
#> 1   379435578    7870139 +42.6381713 -072.2584801
#> 2  1644721618  165889022 +44.3602768 -069.7286675
#> 3  4569506777  195987592 +45.0041529 -068.7830182
#> 4  1150377509  371437958 +41.7335795 -070.3353854
#> 5  1160959555   13052380 +44.2632544 -072.4847537
#> 6   463792807    1998919 +42.8901464 -073.1781568
#> 7   776442799   10872456 +44.4915310 -071.2261426
#> 8  8526630528 1039679917 +42.4047869 -071.1493237
#> 9  1514413834  237609547 +41.2232437 -073.3371782
#> 10  920232630  595603602 +43.8816153 -069.8782047
#>                          geometry
#> 1  MULTIPOLYGON (((-72.28305 4...
#> 2  MULTIPOLYGON (((-69.52822 4...
#> 3  MULTIPOLYGON (((-68.81579 4...
#> 4  MULTIPOLYGON (((-70.03749 4...
#> 5  MULTIPOLYGON (((-72.49324 4...
#> 6  MULTIPOLYGON (((-73.13711 4...
#> 7  MULTIPOLYGON (((-71.16203 4...
#> 8  MULTIPOLYGON (((-71.18679 4...
#> 9  MULTIPOLYGON (((-73.23264 4...
#> 10 MULTIPOLYGON (((-69.8819 43...