Download TIGER shapes for New England City and Town Area Divisions

Download TIGER shapes for New England City and Town Area Divisions

Usage

tt_new_england_city_divisions(year = 2021)

Arguments

year

Integer year. Required. Years < 2012 are not currently supported.

Value

sf data.frame

Examples

tt_new_england_city_divisions()
#> Simple feature collection with 11 features and 12 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -71.85505 ymin: 41.70016 xmax: -70.51636 ymax: 43.0214
#> Geodetic CRS:  NAD83
#> First 10 features:
#>    CNECTAFP NECTAFP NCTADVFP      GEOID
#> 1       715   71650    71634 7165071634
#> 2       715   71650    72104 7165072104
#> 3       715   71650    73104 7165073104
#> 4       715   71650    73604 7165073604
#> 5       715   71650    74204 7165074204
#> 6       715   71650    74804 7165074804
#> 7       715   71650    74854 7165074854
#> 8       715   71650    75404 7165075404
#> 9       715   71650    76524 7165076524
#> 10      715   71650    76684 7165076684
#>                                          NAME
#> 1                 Boston-Cambridge-Newton, MA
#> 2        Brockton-Bridgewater Town-Easton, MA
#> 3                              Framingham, MA
#> 4  Haverhill-Newburyport-Amesbury Town, MA-NH
#> 5  Lawrence-Methuen Town-North Andover, MA-NH
#> 6          Lowell-Billerica-Chelmsford, MA-NH
#> 7                   Lynn-Salem-Marblehead, MA
#> 8                               Nashua, NH-MA
#> 9              Peabody-Beverly-Gloucester, MA
#> 10              Plymouth-Pembroke-Duxbury, MA
#>                                                     NAMELSAD LSAD MTFCC
#> 1                 Boston-Cambridge-Newton, MA NECTA Division   M7 G3220
#> 2        Brockton-Bridgewater Town-Easton, MA NECTA Division   M7 G3220
#> 3                              Framingham, MA NECTA Division   M7 G3220
#> 4  Haverhill-Newburyport-Amesbury Town, MA-NH NECTA Division   M7 G3220
#> 5  Lawrence-Methuen Town-North Andover, MA-NH NECTA Division   M7 G3220
#> 6          Lowell-Billerica-Chelmsford, MA-NH NECTA Division   M7 G3220
#> 7                   Lynn-Salem-Marblehead, MA NECTA Division   M7 G3220
#> 8                               Nashua, NH-MA NECTA Division   M7 G3220
#> 9              Peabody-Beverly-Gloucester, MA NECTA Division   M7 G3220
#> 10              Plymouth-Pembroke-Duxbury, MA NECTA Division   M7 G3220
#>         ALAND    AWATER    INTPTLAT     INTPTLON                       geometry
#> 1  2860347068 425757253 +42.3149834 -071.1675691 MULTIPOLYGON (((-70.98127 4...
#> 2   321545374   7531881 +42.0341015 -071.0157590 MULTIPOLYGON (((-70.97523 4...
#> 3   503220194  23246707 +42.2638445 -071.5106679 MULTIPOLYGON (((-71.37769 4...
#> 4   938908003  56404046 +42.8429020 -071.0224861 MULTIPOLYGON (((-70.94859 4...
#> 5   207887679   9752445 +42.7281251 -071.1632265 MULTIPOLYGON (((-71.11558 4...
#> 6   542144814  20181384 +42.6303724 -071.3662554 MULTIPOLYGON (((-71.33254 4...
#> 7    71248678 119749125 +42.4759260 -070.8857445 MULTIPOLYGON (((-70.83862 4...
#> 8  1244962561  20160392 +42.8520335 -071.5198762 MULTIPOLYGON (((-71.54254 4...
#> 9   466892258 153917480 +42.6160552 -070.8185167 MULTIPOLYGON (((-70.56785 4...
#> 10  591866277 153634043 +41.9416950 -070.6992089 MULTIPOLYGON (((-70.74972 4...