Download TIGER shapes for New England Combined City and Town Areas

Download TIGER shapes for New England Combined City and Town Areas

Usage

tt_new_england_combined_areas(year = 2021)

Arguments

year

Integer year. Required. Years < 2012 are not currently supported.

Value

sf data.frame

Examples

tt_new_england_combined_areas()
#> Simple feature collection with 7 features and 10 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -73.72777 ymin: 40.95094 xmax: -69.33964 ymax: 44.69986
#> Geodetic CRS:  NAD83
#>   CNECTAFP GEOID                                        NAME
#> 1      710   710                      Augusta-Waterville, ME
#> 2      715   715 Boston-Worcester-Providence, MA-RI-NH-CT-ME
#> 3      720   720           Bridgeport-New Haven-Stamford, CT
#> 4      725   725                    Lebanon-Claremont, NH-VT
#> 5      770   770               Pittsfield-North Adams, MA-VT
#> 6      775   775        Portland-Lewiston-South Portland, ME
#> 7      790   790   Springfield-Hartford-East Hartford, MA-CT
#>                                                     NAMELSAD LSAD MTFCC
#> 1                      Augusta-Waterville, ME Combined NECTA   M4 G3200
#> 2 Boston-Worcester-Providence, MA-RI-NH-CT-ME Combined NECTA   M4 G3200
#> 3           Bridgeport-New Haven-Stamford, CT Combined NECTA   M4 G3200
#> 4                    Lebanon-Claremont, NH-VT Combined NECTA   M4 G3200
#> 5               Pittsfield-North Adams, MA-VT Combined NECTA   M4 G3200
#> 6        Portland-Lewiston-South Portland, ME Combined NECTA   M4 G3200
#> 7   Springfield-Hartford-East Hartford, MA-CT Combined NECTA   M4 G3200
#>         ALAND     AWATER    INTPTLAT     INTPTLON
#> 1  2155056550  183916956 +44.4092939 -069.6901717
#> 2 21419136256 3005007067 +42.3307869 -071.3296644
#> 3  3939856867  374600689 +41.3603421 -073.1284227
#> 4  3032350557   58162981 +43.6727226 -072.2484543
#> 5  1524342565   24516395 +42.5337519 -073.1678825
#> 6  6255802933 1536819921 +43.8562034 -070.3192682
#> 7  8709721508  257195724 +42.0359069 -072.6213616
#>                         geometry
#> 1 MULTIPOLYGON (((-69.39599 4...
#> 2 MULTIPOLYGON (((-70.5375 41...
#> 3 MULTIPOLYGON (((-72.98672 4...
#> 4 MULTIPOLYGON (((-72.2695 43...
#> 5 MULTIPOLYGON (((-72.95104 4...
#> 6 MULTIPOLYGON (((-70.06981 4...
#> 7 MULTIPOLYGON (((-72.21911 4...