Download TIGER shapes for New England Combined City and Town Areas

Download TIGER shapes for New England Combined City and Town Areas

Usage

tt_new_england_combined_areas(year = 2021)

Arguments

year

Integer year. Required. 2000 and 2010-2021 are currently supported.

Value

sf data.frame

Examples

tt_new_england_combined_areas()
#> Simple feature collection with 7 features and 10 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -73.72777 ymin: 40.95094 xmax: -69.33964 ymax: 44.69986
#> Geodetic CRS:  NAD83
#> # A tibble: 7 × 11
#>   CNECTAFP GEOID NAME         NAMEL…¹ LSAD  MTFCC   ALAND AWATER INTPT…² INTPT…³
#>   <chr>    <chr> <chr>        <chr>   <chr> <chr>   <dbl>  <dbl> <chr>   <chr>  
#> 1 710      710   Augusta-Wat… August… M4    G3200 2.16e 9 1.84e8 +44.40… -069.6…
#> 2 715      715   Boston-Worc… Boston… M4    G3200 2.14e10 3.01e9 +42.33… -071.3…
#> 3 720      720   Bridgeport-… Bridge… M4    G3200 3.94e 9 3.75e8 +41.36… -073.1…
#> 4 725      725   Lebanon-Cla… Lebano… M4    G3200 3.03e 9 5.82e7 +43.67… -072.2…
#> 5 770      770   Pittsfield-… Pittsf… M4    G3200 1.52e 9 2.45e7 +42.53… -073.1…
#> 6 775      775   Portland-Le… Portla… M4    G3200 6.26e 9 1.54e9 +43.85… -070.3…
#> 7 790      790   Springfield… Spring… M4    G3200 8.71e 9 2.57e8 +42.03… -072.6…
#> # … with 1 more variable: geometry <MULTIPOLYGON [°]>, and abbreviated variable
#> #   names ¹​NAMELSAD, ²​INTPTLAT, ³​INTPTLON
#> # ℹ Use `colnames()` to see all variable names